requirements.txt
# AutoSphere AI Dependencies
# Core AI and ML libraries
langchain_ibm>=0.3.17
ibm_watsonx_ai>=1.3.36
langchain_core>=0.3.75
langgraph>=0.6.6

# HTTP and API libraries
requests>=2.32.5
httpx>=0.27.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# Data processing
pandas>=0.24.2
numpy>=1.26.0

# Jupyter environment (optional, for notebook version)
jupyter>=1.0.0
ipykernel>=6.29.5
ipywidgets>=8.1.7

# Additional utilities
pydantic>=2.7.4
typing-extensions>=4.7


